requirements.txt
fastapi
uvicorn
jinja2
python-multipart
firebase-admin
python-dotenv
httpx
pydantic
pydantic-settings
